src/main/java/pl/sda/structure/tasks/TaskFilterWithCondition.java
package pl.sda.structure.tasks;

import pl.sda.structure.tasks.model.Task;

import java.util.ArrayList;
import java.util.List;

public class TaskFilterWithCondition {

    private final List<Task> allTask;

    public TaskFilterWithCondition(List<Task> allTask) {
        this.allTask = allTask;
    }

    public List<Task> getOnlyWith(TaskCondition condition) {
        List<Task> result = new ArrayList<>();
        //TODO zaimplementuj
        return result;
    }


    public abstract static class TaskCondition {

        abstract boolean meetCriteria(Task task);

    }

}
